feat(condo): DOMA-4405 add downloading pdf files
